backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
boto3>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
